+++ tests/lib.rs
@@ -1,0 +1,12 @@
+extern crate rusty_machine as rm;
+extern crate num as libnum;
+
+pub mod learning {
+    mod lin_reg;
+    mod k_means;
+    mod gp;
+
+    pub mod optim {
+    	mod grad_desc;
+    }
+}