+++ CHANGELOG.md
@@ -2,6 +2,32 @@
 
 This document will be used to keep track of changes made between release versions. I'll do my best to note any breaking changes!
 
+## 0.4.0
+
+This is the biggest release so far. Primarily because the `linalg`
+module has been pulled out into its own crate: [rulinalg](https://github.com/AtheMathmo/rulinalg).
+
+In addition to this there have been a number of improvements to the `linalg`
+and `learning` moduled in this release.
+
+### Breaking Changes
+
+- The `linalg` module pulled out and replaced by reexports of [rulinalg](https://github.com/AtheMathmo/rulinalg).
+All structs are now imported at the `linalg` level, i.e. `linalg::matrix::Matrix` -> `linalg::Matrix`.
+- Decomposition methods now return `Result` instead of panicking on fail.
+- K-Means now has a trait for `Initializer` - which allows generic initialization algorithms.
+
+### Features
+
+- New error handling in both the `linalg` (now rulinalg) and `learning` modules.
+- Bug fixed in eigendecomposition: it can now be used!
+- K-means can now take a generic initialization algorithm.
+
+### Minor Changes
+
+- Optimization and code cleanup in the decomposition methods.
+- Some optimization in the K-Means model.
+
 ## 0.3.3
 
 ### New Contributors